--- pkg/util/plancache/BUILD.bazel
@@ -1,9 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "plancache",
-    srcs = ["util.go"],
-    importpath = "github.com/pingcap/tidb/pkg/util/plancache",
-    visibility = ["//visibility:public"],
-    deps = ["//pkg/types"],
-)
