--- pkg/util/plancache/util.go
@@ -1,36 +1,0 @@
-// Copyright 2023 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package util
-
-import (
-	"github.com/pingcap/tidb/pkg/types"
-)
-
-// PlanCacheMatchOpts store some property used to fetch plan from plan cache
-// The structure set here is to avoid import cycle
-type PlanCacheMatchOpts struct {
-	// paramTypes stores all parameters' FieldType, some different parameters may share same plan
-	ParamTypes []*types.FieldType
-	// limitOffsetAndCount stores all the offset and key parameters extract from limit statement
-	// only used for cache and pick plan with parameters in limit
-	LimitOffsetAndCount []uint64
-	// HasSubQuery indicate whether this query has sub query
-	HasSubQuery bool
-	// StatsVersionHash is the hash value of the statistics version
-	StatsVersionHash uint64
-
-	// Below are some variables that can affect the plan
-	ForeignKeyChecks bool
-}
